built the base model
building the base model is done. there is thill a gpu/cuda/torch error. but it is nothing. currently egnoring it by running the model on cpu. and anohter error is with the whisper model(bcz i am tryna run the model locally so i downloaded the .pt file and how i have to figure out why the file path not working :)